requirements.txt
flask
langchain
langchain-groq
playwright
python-dotenv
markdown
markupsafe